Remove steps table (#17)
Having both the steps table and the steps_by_thread_id table was not essential. The primary benefit of the steps table was to discover the thread_id for a step using just the step_id. The problem with maintaining two sperate views of steps is keeping both tables in sync. Logged batches help here, but writes for steps tend to be large making it difficult to include them in a batch. This is why, previously, we'd sometimes get a "Batch too large" error. Removing the steps table here makes life easier.

Resolves #16